+++ data/src/main/java/com/xquare/data/repository/reports/BugRepositoryImpl.kt
@@ -1,8 +1,8 @@
-package com.xquare.data.repository.bug
+package com.xquare.data.repository.reports
 
 import com.xquare.data.remote.datasource.BugRemoteDataSource
-import com.xquare.domain.entity.bug.BugEntity
-import com.xquare.domain.repository.bug.BugRepository
+import com.xquare.domain.entity.reports.BugEntity
+import com.xquare.domain.repository.reports.BugRepository
 import javax.inject.Inject
 
 class BugRepositoryImpl @Inject constructor(
