+++ domain/src/main/java/com/xquare/domain/entity/reports/BugEntity.kt
@@ -1,4 +1,4 @@
-package com.xquare.domain.entity.bug
+package com.xquare.domain.entity.reports
 
 data class BugEntity(
     val reason: String,
