--- domain/src/main/java/com/xquare/domain/repository/bug/BugRepository.kt
@@ -1,7 +1,0 @@
-package com.xquare.domain.repository.bug
-
-import com.xquare.domain.entity.bug.BugEntity
-
-interface BugRepository {
-    suspend fun uploadBug(bugEntity: BugEntity)
-}
