+++ domain/src/main/java/com/xquare/domain/repository/reports/BugRepository.kt
@@ -1,0 +1,7 @@
+package com.xquare.domain.repository.reports
+
+import com.xquare.domain.entity.reports.BugEntity
+
+interface BugRepository {
+    suspend fun uploadBug(bugEntity: BugEntity)
+}
